Change repository variable to be local and not singleton
When using singleton version of the DatabaseRepository, realm instance that
was created on initialization was tied to the background thread that the
WidgetUpdateWorker operates on. Given that the singleton instance is then
used inside the app, the same realm instance resides which can't be used
on the main thread as a result.