Repackage entire app structure
Previous strategy for packaging was "by type" which means all activities
under activity/ package, all fragments under /fragments etc. As the number
of classes grows, it was harder to find files related to similar
functionality, so current approach is to package "by feature". Data/
package containing classes that deals with local and remote databases,
util/ package to hold utilities and ui/ package that is divided into
child packages with features as names.

Additional note: I am still stuck with the Provider/ package because
changing the location of AppWidgetProvider class would cause existing
widgets on the screen to disappear.